Fix missing macros in liburing.h
Somehow my <linux/io_uring.h> header file is out of sync leading to
multiple macros missing in the liburing.h file. This fixes the problem
since liburing uses it's own io_uring.h file which works, but this
should be investigated if it's just on my machine or for other arch
users as well.